+++ src/site/Components/Contato/contato.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function Contato(){
+    return <section id="preco">
+         <div className="container">
+             <div className="row text-center">
+                <h1>Contato</h1>
+                <p>Página em desenvolvimento</p>
+             </div>
+         </div>
+
+
+        </section>;
+  }
+
+export default Contato;
