+++ src/site/Components/Menu/menu.jsx
@@ -31,13 +31,9 @@ function Menu(){
             <li className="nav-item">
 
             <Link to="/relatorio" className="nav-link" aria-current="page">Transparência</Link>
-   
+
             </li>
 
-            <li className="nav-item">
-            <Link to="/relatorio" className="nav-link" aria-current="page">Relatório</Link>
-            </li>
-    
             <li className="nav-item">
             <Link to="/contato" className="nav-link" aria-current="page">Contatos</Link>
             </li>
