--- src/site/Components/contato.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-function Contato(){
-    return <section id="preco">
-        <div className="container">
-
-            <div className="row text-center">
-
-                    </div>
-                </div>
-
-
-                    </section>;
-  }
-
-export default Contato;
